% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grd.ddist.R
\name{grd.ddist}
\alias{grd.ddist}
\title{grd.ddist}
\usage{
grd.ddist(X, ddistx)
}
\arguments{
\item{X}{excess water value}

\item{ddistx}{distance distribution of each saturation level}
}
\value{
The output is a vector giving the state of each of the level
}
\description{
The function \code{grd.ddist()} computes the states of each saturation level.
Unit is in millimeters.
}
\examples{
\dontrun{
grd.ddist()
}
}
\keyword{groundwater}

