% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grd.ddistx.R
\name{grd.ddistx}
\alias{grd.ddistx}
\title{grd.ddistx}
\usage{
grd.ddistx(Layers, Magkap, nbStepsDelay)
}
\arguments{
\item{Layers}{matrix describing the groundwater; the number of column represent the delay-steps, the row represent the levels (1st is the fastest, the last is the slowest)}

\item{Magkap}{Maximaum capacity}

\item{nbStepsDelay}{delay-steps for each level}
}
\value{
The output is a vector giving the current capacity of each of the level
}
\description{
The function \code{grd.ddistx()} informs about the current capacity of each level (in mm)
}
\examples{
\dontrun{
grd.ddistx()
}
}
\keyword{groundwater}

