% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grd.M.R
\name{grd.M}
\alias{grd.M}
\title{grd.M}
\usage{
grd.M(NoL, Res_prob)
}
\arguments{
\item{NoL}{number of Layers}

\item{Res_prob}{groundwater saturation quantiles of each level}
}
\value{
The output is a scalar
}
\description{
The function \code{grd.M()} computes the Groundwater Storage Capacity (GSC)
}
\examples{
\dontrun{
grd.M()
}
}
\keyword{groundwater}

