% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grd.Magkap.R
\name{grd.Magkap}
\alias{grd.Magkap}
\title{dddGroundwater.Magkap}
\usage{
grd.Magkap(NoL, Res_prob, CapacityUpperLevel)
}
\arguments{
\item{NoL}{number of layer}

\item{Res_prob}{groundwater saturation quantiles for each level}

\item{CapacityUpperLevel}{capacity of the upper level}
}
\value{
The output is a scalar
}
\description{
The function \code{grd.Magkap} computes the maximum capacity for each saturation level
}
\examples{
\dontrun{
grd.Magkap()
}
}
\keyword{groundwater}

