% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grd.Res_prob.R
\name{grd.Res_prob}
\alias{grd.Res_prob}
\title{grd.Res_prob}
\usage{
grd.Res_prob(NoL, UHMAD, MAD, Timeresinsec, area, modelSaturation)
}
\arguments{
\item{NoL}{number of saturation level}

\item{UHMAD}{Unit Hydrograph of the Mean Annual Discharge}

\item{MAD}{value of the Mean Annual Discharge}

\item{Timeresinsec}{time resolution of the process in second (1hour: 3600s, ... etc)}

\item{area}{in squared meters}

\item{modelSaturation}{list of parameters about the saturation}
}
\value{
The output is a groundwater saturation quantile for each saturation level
}
\description{
The function \code{grd.Res_Prob()} computes the groundwater saturation quantiles
}
\examples{
\dontrun{
grd.Res_prob()
}
}
\keyword{groundwater}

