% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grd.S.R
\name{grd.S}
\alias{grd.S}
\title{grd.S}
\usage{
grd.S(ddistx, X, ddist)
}
\arguments{
\item{ddistx}{distance distribution of each saturation level}

\item{X}{excess water, scalar}

\item{ddist}{states (in mm) of each saturation level}
}
\value{
The output is a scalar
}
\description{
The function \code{grd.S()} computes the deficit S for all sub surface layers (not the overland flow layer)
}
\details{
It follows the expression \code{dS/dt = -dD/dt} from  Skaugen,Peerebom and Nilsson(2015)
}
\examples{
\dontrun{
grd.S()
}
}
\keyword{groundwater}

