% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.ddistAll.R
\name{init.ddistAll}
\alias{init.ddistAll}
\title{ddistAll initialization
The funtion initializes the main features of the groundwater:
- The maximum capacity of each staturation level,
- the Groundwater Storage Capacity (GSC),
- The saturation layers}
\usage{
init.ddistAll(method = NULL, path = NULL, S = NULL, ddistx = NULL,
  ddist = NULL)
}
\arguments{
\item{method}{method for the initialization, "load", "manual"}

\item{path}{directory where to get the files}

\item{S}{deficit}

\item{ddistx}{current capacity of each level (in mm)}

\item{ddist}{states of each saturation level}
}
\description{
ddistAll initialization
The funtion initializes the main features of the groundwater:
- The maximum capacity of each staturation level,
- the Groundwater Storage Capacity (GSC),
- The saturation layers
}
\examples{
\dontrun{
init.ddistAll())
}
}
\keyword{groundwater}

