% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.groundwater.R
\name{init.groundwater}
\alias{init.groundwater}
\title{Groundwater initialization
The funtion initializes the main features of the groundwater:
- The maximum capacity of each staturation level,
- the Groundwater Storage Capacity (GSC),
- The saturation layers}
\usage{
init.groundwater(method = NULL, path = NULL, Magkap = NULL, M = NULL,
  Layers = NULL, Timeresinsec = NULL, UHMAD = NULL, MAD = NULL,
  modelArea = NULL, modelSaturation = NULL, modelLayer = NULL)
}
\arguments{
\item{method}{method for the initialization, "load", "source", "manual"}

\item{path}{directory where to get the files}

\item{Magkap}{Magkap}

\item{M}{Groundwater Storage Capacity (GSC)}

\item{Layers}{saturation layers}

\item{Timeresinsec}{time resolution of the process in second (1hour: 3600s, ... etc)}

\item{UHMAD}{Unit Hydrograph of the Mean Annual Discharge}

\item{MAD}{value of the Mean Annual Discharge}

\item{modelArea}{list(totarea,slopesriverarea,nobognoglacarea,bogarea)}

\item{modelSaturation}{list of parameters about the saturation
list(gtcel,CapacityUpperLevel,mLam,varLam,distr)}

\item{modelLayer}{list of parameters about the Layers
list(maxL,speed,nbStepsDelay,z,distr,param,NoL)}
}
\description{
Groundwater initialization
The funtion initializes the main features of the groundwater:
- The maximum capacity of each staturation level,
- the Groundwater Storage Capacity (GSC),
- The saturation layers
}
\examples{
\dontrun{
init.groundwater)
}
}
\keyword{groundwater}

