% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{init
The funtion initialize the main features of the groundwater:
- The maximum capacity of each staturation level,
- the Groundwater Storage Capacity (GSC),
- The saturation layers}
\usage{
init(Timeresinsec, UHMAD, MAD, area, modelSaturation, modelLayer)
}
\arguments{
\item{Timeresinsec}{time resolution of the process in second (1hour: 3600s, ... etc)}

\item{UHMAD}{Unit Hydrograph of the Mean Annual Discharge}

\item{MAD}{value of the Mean Annual Discharge}

\item{area}{in squared meters}

\item{modelSaturation}{list of parameters about the saturation
list(gtcel=,CapacityUpperLevel=b,mLam=c,varLam=d,distr="qgamma")}

\item{modelLayer}{list of parameters about the Layers
list(maxL=a,speed=b,nbStepsDelay=c,z=d,distr="dexp",param=c(e), NoL=f)}
}
\description{
init
The funtion initialize the main features of the groundwater:
- The maximum capacity of each staturation level,
- the Groundwater Storage Capacity (GSC),
- The saturation layers
}
\examples{
\dontrun{
init()
}
}
\keyword{groundwater}

