% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saturationSum.R
\name{saturationSum}
\alias{saturationSum}
\title{saturationSum}
\usage{
saturationSum(UHMAD, MAD, Timeresinsec, area)
}
\arguments{
\item{UHMAD}{Unit Hydrograph of the Mean Annual Discharge}

\item{MAD}{value of the Mean Annual Discharge}

\item{Timeresinsec}{time resolution of the process in second (1hour: 3600s, ... etc)}

\item{area}{in squared meters}
}
\description{
the function \code{saturationSum()} computes the saturation sum
}
\examples{
\dontrun{
saturationSum()
}
}
\keyword{groundwater}

