% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stateX.ddistAll.R
\name{stateX.ddistAll}
\alias{stateX.ddistAll}
\title{stateX.ddistAll}
\usage{
stateX.ddistAll(Layers, Magkap, nbStepsDelay, X)
}
\arguments{
\item{Layers}{matrix describing the groundwater; the number of column represent the delay-steps, the row represent the levels (1st is the fastest, the last is the slowest)}

\item{Magkap}{Maximaum capacity}

\item{nbStepsDelay}{delay-steps for each level}

\item{X}{excess water value}
}
\value{
The output is a list providing the three main features for groundwater( ddistx, ddist, S)
}
\description{
The function \code{stateX.ddistAll()} gathers three processes:
- it informs on current capacity for each level in mm.
- it computes the states (in mm) of each saturation level
- it computes the deficit S (for all sub surface layers, NOT overland flow layer)
}
\examples{
\dontrun{
ddistAll()
}
}
\keyword{groundwater}

