% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stateX.groundwater.R
\name{stateX.groundwater}
\alias{stateX.groundwater}
\title{stateX.layers
The function \code{stateX.groundwater} updates the saturation levels.
It computes the runoff event that occurs in each of layer, shift the layer by one timestep and then update the layer value with the runoff value}
\usage{
stateX.groundwater(NoL, Layers, ddist, X, layerUH, nbStepsDelay)
}
\arguments{
\item{NoL}{number of level zone}

\item{Layers}{matrix describing the groundwater; the number of column represent the delay-steps, the row represent the levels (1st is the fastest, the last is the slowest)}

\item{ddist}{states (in mm) of each saturation level}

\item{X}{xcess water}

\item{layerUH}{Unit Hydrograph of the layer}

\item{nbStepsDelay}{delay-steps of each layer}
}
\value{
The output is a matrix describing the groundwater.
 the number of column represent the delay-steps, the row represent the levels (1st is the fastest, the last is the slowest)
}
\description{
stateX.layers
The function \code{stateX.groundwater} updates the saturation levels.
It computes the runoff event that occurs in each of layer, shift the layer by one timestep and then update the layer value with the runoff value
}
\examples{
\dontrun{
stateX.layers()
}
}
\keyword{groundwater}

